refactor: raname `vp_format.py` in `vp_formats.py`
